HMS-5296: Rename adminTasks folder > Admin
HMS-5296: fix failing tests
