+++ src/Pages/Repositories/AdminTaskTable/AdminTaskTable.test.tsx
@@ -1,10 +1,10 @@
 import { render } from '@testing-library/react';
 import { ReactQueryTestWrapper, defaultIntrospectTask } from 'testingHelpers';
 import AdminTaskTable from './AdminTaskTable';
-import { useAdminTaskListQuery } from 'services/AdminTasks/AdminTaskQueries';
+import { useAdminTaskListQuery } from 'services/Admin/AdminTaskQueries';
 import { formatDateDDMMMYYYY } from 'helpers';
 
-jest.mock('services/AdminTasks/AdminTaskQueries', () => ({
+jest.mock('services/Admin/AdminTaskQueries', () => ({
   useAdminTaskListQuery: jest.fn(),
   useFetchAdminTaskQuery: () => ({ fetchAdminTask: () => undefined, isLoading: false }),
 }));
